formatting and documentation

+++ gr/auth/ee/lcs/evaluators/AccuracyEvaluator.java
@@ -3,19 +3,21 @@
  */
 package gr.auth.ee.lcs.evaluators;
 
+import gr.auth.ee.lcs.classifiers.ClassifierSet;
+import gr.auth.ee.lcs.data.ClassifierTransformBridge;
+import gr.auth.ee.lcs.data.IEvaluator;
+
 import java.io.FileReader;
 import java.io.IOException;
 import java.util.Arrays;
 
 import weka.core.Instances;
-import gr.auth.ee.lcs.classifiers.ClassifierSet;
-import gr.auth.ee.lcs.data.ClassifierTransformBridge;
-import gr.auth.ee.lcs.data.IEvaluator;
 
 /**
  * An evaluator to evaluate on the accuracy of the classifiers
+ * 
  * @author Miltos Allamanis
- *
+ * 
  */
 public class AccuracyEvaluator implements IEvaluator {
 
@@ -28,26 +30,48 @@ public class AccuracyEvaluator implements IEvaluator {
 	 * A boolean indicating if the evaluator is going to print the results.
 	 */
 	private final boolean printResults;
-	
+
+	/**
+	 * Constructor for creating evaluator with a Weka instance set
+	 * 
+	 * @param instances
+	 *            the instances to be used
+	 * @param print
+	 *            true to print results to stdout
+	 */
 	public AccuracyEvaluator(final Instances instances, final boolean print) {
 		instanceSet = instances;
 		printResults = print;
 	}
 
+	/**
+	 * Constructor for creating evaluator with .arff file
+	 * 
+	 * @param arffFileName
+	 *            the arff file
+	 * @param print
+	 *            true to print output to stdout
+	 * @throws IOException
+	 *             when file is not found
+	 */
 	public AccuracyEvaluator(final String arffFileName, final boolean print)
 			throws IOException {
 		printResults = print;
 		FileReader reader = new FileReader(arffFileName);
 		instanceSet = new Instances(reader);
 	}
-	
-	/* (non-Javadoc)
-	 * @see gr.auth.ee.lcs.data.IEvaluator#evaluateSet(gr.auth.ee.lcs.classifiers.ClassifierSet)
+
+	/*
+	 * (non-Javadoc)
+	 * 
+	 * @see
+	 * gr.auth.ee.lcs.data.IEvaluator#evaluateSet(gr.auth.ee.lcs.classifiers
+	 * .ClassifierSet)
 	 */
 	@Override
 	public double evaluateSet(ClassifierSet classifiers) {
 		final ClassifierTransformBridge bridge = ClassifierTransformBridge
-		.getInstance();
+				.getInstance();
 		double sumOfAccuracies = 0;
 		double sumOfRecall = 0;
 		for (int i = 0; i < instanceSet.numInstances(); i++) {
@@ -59,7 +83,7 @@ public class AccuracyEvaluator implements IEvaluator {
 			}
 			final int[] classes = bridge.classify(classifiers, instance);
 			final int[] classification = bridge.getDataInstanceLabels(instance);
-		
+
 			// Find symetric differences
 			Arrays.sort(classes);
 			Arrays.sort(classification);
@@ -67,20 +91,22 @@ public class AccuracyEvaluator implements IEvaluator {
 				if (Arrays.binarySearch(classification, classes[j]) < 0) {
 					unionOfLabels++;
 				} else {
-					intersectionOfLabels ++;
-					unionOfLabels ++;
+					intersectionOfLabels++;
+					unionOfLabels++;
 				}
 			}
 			for (int j = 0; j < classification.length; j++) {
 				if (Arrays.binarySearch(classes, classification[j]) < 0)
 					unionOfLabels++;
 			}
-			sumOfAccuracies += ((double)intersectionOfLabels) / ((double)unionOfLabels);
-			sumOfRecall += ((double)intersectionOfLabels) / ((double) classification.length);
+			sumOfAccuracies += ((double) intersectionOfLabels)
+					/ ((double) unionOfLabels);
+			sumOfRecall += ((double) intersectionOfLabels)
+					/ ((double) classification.length);
 		}
-		final double accuracy = sumOfAccuracies / ((double)instanceSet.numInstances());
-		final double recall =  sumOfRecall / ((double)instanceSet.numInstances());
-		if (printResults){
+		final double accuracy = sumOfAccuracies / (instanceSet.numInstances());
+		final double recall = sumOfRecall / (instanceSet.numInstances());
+		if (printResults) {
 			System.out.println("Accuracy: " + accuracy);
 			System.out.println("Recall: " + recall);
 		}
